Fixing up the DB more.
I also switched it so that the tokens would be String based (UUID +
other data). This will make it less obvjois whats going on behind the
scenes, and if we ever need to change the API hasn't changed at all.

Also added the scema mapper so that i'm able to use different schema's
for dev and prod